Refactored garbage collection: split `moveObjects` and added `updatePtrHolder` for pointer management, changed `forwarded` to 64-bit, and introduced `OP_ARR_INIT`. Made-by: philipp <philovic@outlook.de>

+++ src/vm/heap.h
@@ -17,7 +17,7 @@ namespace sprout::heap {
         uint32_t size;
         uint16_t type;
         uint16_t flags;
-        uint32_t forwarded;
+        uint64_t forwarded;
     };
 
     enum objType {
